\name{AIFFtools-package}
\alias{AIFFtools-package}
\alias{AIFFtools}
\docType{package}
\title{
\packageTitle{AIFFtools}
}
\description{
\packageDescription{AIFFtools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AIFFtools}
\packageIndices{AIFFtools}

}
\author{
\packageAuthor{AIFFtools}

Maintainer: \packageMaintainer{AIFFtools}
}

\keyword{ package }

